Make code work without TiGL/TiXi and ensure FlexiFlightVis works
The variables "aircraft" and "structure" can now be
loaded from the file "aircraft_structure.mat" in
case no TiGL/TiXI installation can be found.
Moreover, FlexiFlightVis did not work properly
in some scripts where the Fast Restart option of
Simulink is used. This was due to an initialize
problem of the Simulink block "Send to
FlexiFlightVis". The problem was solved by updating
the Simulink model before the Fast Restart option
is activated.